Releasing version 12 4 0
Squashed commit of the following:

commit b3e6b1d59a549cdf124cd6c2eea0153688987bc1
Author: dshelbyo <derek.shelby@oracle.com>
Date:   Fri Nov 8 21:17:13 2019 +0000

    Releasing version 12 4 0

+++ objectstorage/transfer/file_uploader_test.go
@@ -116,7 +116,7 @@ func TestUploadFileMultiparts(t *testing.T) {
 	for _, testData := range testDataSet {
 		fake := fake{upLoadID: testData.uploadID, failedPartNumbers: testData.failedPartNumbers, numberOfCommitedParts: common.Int(0), numberOfUploadedParts: common.Int(0)}
 		fileUpload.multipartUploader = &fake
-		filePath, _ := helpers.WriteTempFileOfSize(int64(fileSize))
+		filePath, _ := helpers.WriteTempFileOfSize(fileSize)
 		request := UploadFileRequest{
 			UploadRequest: UploadRequest{PartSize: common.Int64(partSize)},
 			FilePath:      filePath,
